Bad things are going to happen if a negative sequence is passed to COM_BlockSequenceCRCByte (can we say memory access violation?). A sequence number should never be < 0, but we cast it to unsigned int now just to be sure

+++ trunk/zquake/common.c
@@ -1972,7 +1972,7 @@ byte COM_BlockSequenceCRCByte (byte *base, int length, int sequence)
 	byte	*p;
 	byte chkb[60 + 4];
 
-	p = chktbl + (sequence % (sizeof(chktbl) - 4));
+	p = chktbl + ((unsigned int)sequence % (sizeof(chktbl) - 4));
 
 	if (length > 60)
 		length = 60;
